Updates localnet setup and documentation
Refactors the localnet setup to use more consistent naming and
configuration.

Updates the validator startup script path in the documentation
to reflect the change to `validator-localnet.sh`.

Comments out validator key mappings in docker-compose to avoid
key duplication issues.

Updates the validator name and cluster environment to be more
appropriate for a localnet setup.