52 fix setup script for running validator (#53)
* Remove limitation of installing only in hot-spare hosts

A host running a primary identity can also be setup under new conditions. It is responsibility of the operator to know what its being done.

* Include jito_relayer_cohosted role main instead of separate tasks

* Remove unneeded pre-checks | refactor task names

* Remove unneeded variable

Variable jito_relayer_verify was removed since we will always verify jito relayer as long as a co-hosted setup is executed

* Remove relayer build dir before installing the new one

Refactored the workflow to build the relayer based on:
- Check which version is installed and require install if it is not installed or installed version does not match the operator's specified version
- Refactored the way version number is parsed after install as verification method

* Refactored task names for install tasks

* Always run verification